chore: moving functions to where I expect them to be
